Migration updated

+++ app/database/migrations/2014_05_14_190329_create_images_table.php
@@ -15,7 +15,7 @@ class CreateImagesTable extends Migration {
 		Schema::create('images', function(Blueprint $table) {
 			$table->increments('id');
 			$table->string('name');
-			$table->integer('size');
+			$table->bigInteger('size');
 			$table->string('description');
 			$table->string('url');
 			$table->string('thumbnailUrl');
